django-backend/requirements.txt
Django==4.2.6
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
locust~=2.29.0
